+++ Tests/TreasuryBankJournal/New-BankJournalExportCounts.Tests.ps1
@@ -5,9 +5,9 @@
 Describe "New-BankJournalExportCounts" {
 	BeforeAll { . "$PSScriptRoot/../BeforeAll.ps1" }
 
-	It "should return a new bank journal export request" {
-		$exportRequest = New-AgicalExportCounts -CurrentBankJournalEntriesCountInYear 666 -CurrentBankJournalsCountInYear 123
-		Should-Be 666 $exportRequest.CurrentBankJournalEntriesCountInYear
-		Should-Be 123 $exportRequest.CurrentBankJournalsCountInYear
+	It "should return new bank journal export counts" {
+		$exportCounts = New-AgicalExportCounts -CurrentBankJournalEntriesCountInYear 666 -CurrentBankJournalsCountInYear 123
+		Should-Be 666 $exportCounts.CurrentBankJournalEntriesCountInYear
+		Should-Be 123 $exportCounts.CurrentBankJournalsCountInYear
 	}
 }